Renamed fill files to close
