+++ xlb/operator/boundary_masker/aabb_close.py
